inventory: merge to trx
