+++ src/libtrx/include/libtrx/game/stats/common.h
@@ -20,6 +20,7 @@ extern uint32_t Stats_GetMaxSecretFlags(void);
 void Stats_UpdateSecrets(LEVEL_STATS *stats);
 extern void Stats_MarkSecretCollected(const ITEM *item);
 extern bool Stats_CheckAllSecretsCollected(GF_LEVEL_TYPE level_type);
+extern bool Stats_CheckAllLevelSecretsCollected(void);
 
 void Stats_AddMedipacksUsed(double medipack_value);
 void Stats_AddDeath(void);
