+++ src/tr1/game/savegame/savegame_bson.c
@@ -1,4 +1,3 @@
-#include "game/inventory.h"
 #include "game/savegame.h"
 #include "game/stats.h"
 
@@ -10,6 +9,7 @@
 #include <libtrx/game/effects.h>
 #include <libtrx/game/game.h>
 #include <libtrx/game/game_flow.h>
+#include <libtrx/game/inventory.h>
 #include <libtrx/game/lara.h>
 #include <libtrx/game/music.h>
 #include <libtrx/game/objects/general/lift.h>
