--- src/tr2/game/inventory.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/inventory.h>
