+++ src/tr2/game/savegame/savegame_bson.c
@@ -1,4 +1,3 @@
-#include "game/inventory.h"
 #include "game/savegame.h"
 
 #include <libtrx/bson.h>
@@ -9,6 +8,7 @@
 #include <libtrx/game/effects.h>
 #include <libtrx/game/game.h>
 #include <libtrx/game/game_flow.h>
+#include <libtrx/game/inventory.h>
 #include <libtrx/game/lara.h>
 #include <libtrx/game/music.h>
 #include <libtrx/game/objects/general/lift.h>
