+++ src/tr2/game/savegame/savegame_legacy.c
@@ -1,5 +1,4 @@
 #include "game/game_string.h"
-#include "game/inventory.h"
 #include "game/savegame.h"
 
 #include <libtrx/debug.h>
@@ -7,6 +6,7 @@
 #include <libtrx/game/carrier.h>
 #include <libtrx/game/game.h>
 #include <libtrx/game/game_flow.h>
+#include <libtrx/game/inventory.h>
 #include <libtrx/game/lara.h>
 #include <libtrx/game/music.h>
 #include <libtrx/game/objects/general/lift.h>
