+++ src/tr2/game/stats.h
@@ -5,7 +5,5 @@
 
 FINAL_STATS Stats_ComputeFinalStats(GF_LEVEL_TYPE level_type);
 
-int32_t Stats_GetMaxSecrets(void);
-bool Stats_CheckAllLevelSecretsCollected(void);
 uint32_t Stats_ReserveSecretBit(OBJECT_ID object_id);
 OBJECT_ID Stats_GetSecretObject(int32_t secret_idx);
