+++ src/tr2/meson.build
@@ -93,7 +93,6 @@ sources = [
   'game/demo.c',
   'game/game_string.c',
   'game/gun/gun_rifle.c',
-  'game/inventory.c',
   'game/level.c',
   'game/objects/common.c',
   'game/objects/general/harpoon_bolt.c',
